@@ -2,18 +2,17 @@
     "name": "orchesty/php-sdk",
     "description": "PHP sdk for Orchesty",
     "license": "Apache-2.0",
-    "version": "2.3.5",
+    "version": "3.0.0",
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "ext-fileinfo": "*",
         "ext-json": "*",
         "ext-mongodb": "*",
         "ext-oauth": "*",
-        "doctrine/mongodb-odm-bundle": "^4.5",
-        "hanaboso/commons-bundle": "^2.7",
+        "hanaboso/commons-bundle": "^2.8",
         "hanaboso/mongo-data-grid": "^1.6",
         "league/oauth2-client": "^2.6",
-        "phpoffice/phpspreadsheet": "^1.25",
+        "phpoffice/phpspreadsheet": "^1.27",
         "symfony/monolog-bundle": "^v3.8",
         "symfony/yaml": "^v6.2"
     },
