@@ -3,8 +3,11 @@
 namespace PipesPhpSdkTests\Integration\Application\Manager\Webhook;
 
 use Closure;
-use Doctrine\Persistence\ObjectRepository;
 use Exception;
+use GuzzleHttp\Exception\GuzzleException;
+use GuzzleHttp\Psr7\Response;
+use Hanaboso\CommonsBundle\Transport\Curl\CurlException;
+use Hanaboso\CommonsBundle\Transport\Curl\CurlManager;
 use Hanaboso\CommonsBundle\Transport\Curl\Dto\ResponseDto;
 use Hanaboso\CommonsBundle\Transport\CurlManagerInterface;
 use Hanaboso\PipesPhpSdk\Application\Document\ApplicationInstall;
@@ -12,16 +15,26 @@ use Hanaboso\PipesPhpSdk\Application\Document\Webhook;
 use Hanaboso\PipesPhpSdk\Application\Exception\ApplicationInstallException;
 use Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookManager;
 use Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookSubscription;
+use Hanaboso\PipesPhpSdk\Application\Repository\ApplicationInstallRepository;
 use Hanaboso\PipesPhpSdk\Application\Repository\WebhookRepository;
-use PipesPhpSdkTests\DatabaseTestCaseAbstract;
+use Hanaboso\Utils\Exception\DateTimeException;
+use Hanaboso\Utils\String\Json;
+use PipesPhpSdkTests\KernelTestCaseAbstract;
+use PipesPhpSdkTests\MockServer\Mock;
+use PipesPhpSdkTests\MockServer\MockServer;
 
 /**
  * Class WebhookManagerTest
  *
  * @package PipesPhpSdkTests\Integration\Application\Manager\Webhook
  */
-final class WebhookManagerTest extends DatabaseTestCaseAbstract
+final class WebhookManagerTest extends KernelTestCaseAbstract
 {
+
+    /**
+     * @var MockServer $mockServer
+     */
+    private MockServer $mockServer;
 
     /**
      * @var WebhookApplication
@@ -29,9 +42,9 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
     private WebhookApplication $application;
 
     /**
-     * @var ObjectRepository<Webhook>&WebhookRepository
+     * @var WebhookRepository $repository
      */
-    private $repository;
+    private WebhookRepository $repository;
 
     /**
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookManager::subscribeWebhooks
@@ -40,23 +53,94 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookSubscription::getTopology
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookSubscription::getNode
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookSubscription::getName
+     * @covers \Hanaboso\PipesPhpSdk\Application\Document\Webhook::fromArray
+     * @covers \Hanaboso\PipesPhpSdk\Application\Document\Webhook::toArray
+     * @covers \Hanaboso\PipesPhpSdk\Application\Document\Webhook
      *
+     * @return void
+     * @throws ApplicationInstallException
+     * @throws GuzzleException
+     * @throws CurlException
+     * @throws DateTimeException
      * @throws Exception
      */
     public function testSubscribeAndUnsubscribe(): void
     {
-        $this->dm->persist((new ApplicationInstall())->setUser('User')->setKey('webhook'));
-        $this->dm->flush();
-
+        $this->mockServer = new MockServer();
+        self::getContainer()->set('hbpf.worker-api', $this->mockServer);
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/ApplicationInstall?filter={"names":["webhook"],"users":["User"]}',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(
+                    200,
+                    [],
+                    Json::encode((new ApplicationInstall())->setUser('User')->setKey('webhook')->toArray()),
+                ),
+            ),
+        );
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook',
+                Json::decode(
+                    '[{"id":null,"user":"User","application":"webhook","created":"2023-02-13 11:18:55","name":"name","webhookId":"id","node":"name","token":"a344c694874a1ebb8fb0881714c2c424b3e5fbd895cded820c","topology":"topology"}]',
+                ),
+                CurlManager::METHOD_POST,
+                new Response(200, [], '[]'),
+                ['created' => '2023-02-13 11:18:55', 'token' => 'a344c694874a1ebb8fb0881714c2c424b3e5fbd895cded820c'],
+            ),
+        );
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(
+                    200,
+                    [],
+                    Json::encode(
+                        (new Webhook(
+                            [
+                                'user'        => 'User',
+                                'token'       => 'a344c694874a1ebb8fb0881714c2c424b3e5fbd895cded820c',
+                                'node'        => 'node',
+                                'topology'    => 'topology',
+                                'webhook'     => 'webhook',
+                                'application' => 'webhook',
+                                'webhookId'   => 'id',
+                            ],
+                        ))->toArray(),
+                    ),
+                ),
+                ['created' => '2023-02-13 11:18:55', 'token' => 'a344c694874a1ebb8fb0881714c2c424b3e5fbd895cded820c'],
+            ),
+        );
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook?filter={"applications":["webhook"],"user_uds":["User"]}',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(200, [], '[]'),
+            ),
+        );
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(200, [], '[]'),
+            ),
+        );
+        $this->privateSetUp();
         $this->getService(static fn(): ResponseDto => new ResponseDto(200, 'OK', '{"id":"id"}', []))
             ->subscribeWebhooks($this->application, 'User');
-        $this->dm->clear();
 
         /** @var Webhook[] $webhooks */
-        $webhooks = $this->repository->findAll();
+        $webhooks = $this->repository->findMany();
         self::assertCount(1, $webhooks);
         self::assertEquals('User', $webhooks[0]->getUser());
-        self::assertEquals(50, strlen($webhooks[0]->getToken()));
+        self::assertEquals(50, strlen($webhooks[0]->getToken() ?? ''));
         self::assertEquals('node', $webhooks[0]->getNode());
         self::assertEquals('topology', $webhooks[0]->getTopology());
         self::assertEquals('webhook', $webhooks[0]->getApplication());
@@ -66,7 +150,7 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
         $this->getService(static fn(): ResponseDto => new ResponseDto(200, 'OK', '{"success":true}', []))
             ->unsubscribeWebhooks($this->application, 'User');
 
-        self::assertCount(0, $this->repository->findAll());
+        self::assertCount(0, $this->repository->findMany());
     }
 
     /**
@@ -74,21 +158,72 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookManager::unsubscribeWebhooks
      *
      * @throws Exception
+     * @throws GuzzleException
      */
     public function testSubscribeAndUnsubscribeFailed(): void
     {
-        $this->dm->persist((new ApplicationInstall())->setUser('User')->setKey('webhook'));
-        $this->dm->flush();
-
+        $this->mockServer = new MockServer();
+        self::getContainer()->set('hbpf.worker-api', $this->mockServer);
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/ApplicationInstall?filter={"names":["webhook"],"users":["User"]}',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(200, [], '[{}]'),
+            ),
+        );
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook',
+                Json::decode(
+                    '[{"id":null,"user":"User","application":"webhook","created":"2023-02-13 11:55:26","name":"name","webhookId":"id","node":"name","token":"7d2fe1873b77049267371062a784c4923b65a6e4a3cf549294","topology":"topology"}]',
+                ),
+                CurlManager::METHOD_POST,
+                new Response(
+                    200,
+                    [],
+                    Json::encode((new ApplicationInstall())->setUser('User')->setKey('webhook')->toArray()),
+                ),
+                ['created' => '2023-02-13 11:55:26', 'token' => '7d2fe1873b77049267371062a784c4923b65a6e4a3cf549294'],
+            ),
+        );
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook?filter={"applications":["webhook"],"user_uds":["User"]}',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(200, [], '[{}]'),
+            ),
+        );
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/ApplicationInstall?filter={"names":["webhook"],"users":["User"]}',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(200, [], '[{}]'),
+            ),
+        );
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(
+                    200,
+                    [],
+                    '[{"user":"User","node":"node","topology":"topology","application":"webhook","webhookId":"id","unsubscribeFailed":true}]',
+                ),
+            ),
+        );
+        $this->privateSetUp();
         $this->getService(static fn(): ResponseDto => new ResponseDto(200, 'OK', '{"id":"id"}', []))
             ->subscribeWebhooks($this->application, 'User');
-        $this->dm->clear();
 
         $this->getService(static fn(): ResponseDto => new ResponseDto(200, 'OK', '{"success":false}', []))
             ->unsubscribeWebhooks($this->application, 'User');
 
         /** @var Webhook[] $webhooks */
-        $webhooks = $this->repository->findAll();
+        $webhooks = $this->repository->findMany();
         self::assertCount(1, $webhooks);
         self::assertEquals('User', $webhooks[0]->getUser());
         self::assertEquals('node', $webhooks[0]->getNode());
@@ -102,9 +237,21 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookManager::subscribeWebhooks
      *
      * @throws Exception
+     * @throws GuzzleException
      */
     public function testSubscribeAndUnsubscribeNoApplication(): void
     {
+        $this->mockServer = new MockServer();
+        self::getContainer()->set('hbpf.worker-api', $this->mockServer);
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/ApplicationInstall?filter={"names":["webhook"],"users":["User"]}',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(200, [], '[]'),
+            ),
+        );
+        $this->privateSetUp();
         self::expectException(ApplicationInstallException::class);
         self::expectExceptionCode(ApplicationInstallException::APP_WAS_NOT_FOUND);
 
@@ -117,16 +264,23 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookSubscription::getTopology
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookSubscription::getName
      *
+     * @return void
+     * @throws GuzzleException
      * @throws Exception
      */
     public function testGetWebhooks(): void
     {
-        $webhook = (new Webhook())
-            ->setUser('user')
-            ->setApplication('webhook')
-            ->setName('name')
-            ->setTopology('1');
-        $this->pfd($webhook);
+        $this->mockServer = new MockServer();
+        self::getContainer()->set('hbpf.worker-api', $this->mockServer);
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook?filter={"applications":["webhook"],"user_uds":["user"]}',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(200, [], '[{"name":"name","default":true,"enabled":false,"topology":"1"}]'),
+            ),
+        );
+        $this->privateSetUp();
 
         $result = $this->getService(static fn(): ResponseDto => new ResponseDto(200, 'OK', '{"id":"id"}', []))
             ->getWebhooks($this->application, 'user');
@@ -147,12 +301,12 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookSubscription::getParameters
      *
      * @throws Exception
+     * @throws GuzzleException
      */
     public function testSubscribeWebhooks(): void
     {
+        $this->privateSetUp();
         $params = (new WebhookSubscription('name', 'node', 'topo', []))->getParameters();
-        $this->dm->persist((new ApplicationInstall())->setUser('user')->setKey('webhook'));
-        $this->dm->flush();
 
         $this->getService(static fn(): ResponseDto => new ResponseDto(200, 'OK', '{"id":"id"}', []))
             ->subscribeWebhooks($this->application, 'user', ['name' => 'testName']);
@@ -161,22 +315,25 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
     }
 
     /**
+     * @covers \Hanaboso\PipesPhpSdk\Application\Repository\WebhookFilter
      * @covers \Hanaboso\PipesPhpSdk\Application\Manager\Webhook\WebhookManager::unsubscribeWebhooks
      *
      * @throws Exception
+     * @throws GuzzleException
      */
     public function testUnsubscribeWebhooks(): void
     {
-        $this->dm->persist((new ApplicationInstall())->setUser('user')->setKey('webhook'));
-        $this->dm->flush();
-
-        $webhook = (new Webhook())
-            ->setUser('user')
-            ->setApplication('webhook')
-            ->setName('name')
-            ->setTopology('1');
-        $this->pfd($webhook);
-
+        $this->mockServer = new MockServer();
+        self::getContainer()->set('hbpf.worker-api', $this->mockServer);
+        $this->mockServer->addMock(
+            new Mock(
+                '/document/Webhook?filter={"applications":["webhook"],"user_uds":["user"]}',
+                NULL,
+                CurlManager::METHOD_GET,
+                new Response(200, [], '[]'),
+            ),
+        );
+        $this->privateSetUp();
         $this
             ->getService(static fn(): ResponseDto => new ResponseDto(200, 'OK', '{"id":"id"}', []))
             ->unsubscribeWebhooks($this->application, 'user', ['topology' => 'testTopo']);
@@ -187,12 +344,11 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
     /**
      * @throws Exception
      */
-    protected function setUp(): void
+    protected function privateSetUp(): void
     {
-        parent::setUp();
-
         $this->application = self::getContainer()->get('hbpf.application.webhook');
-        $this->repository  = $this->dm->getRepository(Webhook::class);
+
+        $this->repository = self::getContainer()->get('hbpf.webhook.repository');
     }
 
     /**
@@ -206,7 +362,10 @@ final class WebhookManagerTest extends DatabaseTestCaseAbstract
         $manager = self::createMock(CurlManagerInterface::class);
         $manager->expects(self::any())->method('send')->willReturnCallback($closure);
 
-        return new WebhookManager($this->dm, $manager, 'https://example.com');
+        /** @var ApplicationInstallRepository $applicationInstallRepository */
+        $applicationInstallRepository = self::getContainer()->get('hbpf.application_install.repository');
+
+        return new WebhookManager($applicationInstallRepository, $this->repository, $manager, 'https://example.com');
     }
 
 }
