@@ -36,7 +36,7 @@ final class NullApplication extends BasicApplicationAbstract implements WebhookA
      */
     public function getAuthorizationType(): string
     {
-        return AuthorizationTypeEnum::BASIC;
+        return AuthorizationTypeEnum::BASIC->value;
     }
 
     /**
@@ -194,7 +194,7 @@ final class NullApplication extends BasicApplicationAbstract implements WebhookA
      */
     public function getApplicationType(): string
     {
-        return ApplicationTypeEnum::WEBHOOK;
+        return ApplicationTypeEnum::WEBHOOK->value;
     }
 
     /**
