@@ -50,7 +50,7 @@ final class WebhookApplication extends ApplicationAbstract implements WebhookApp
      */
     public function getAuthorizationType(): string
     {
-        return AuthorizationTypeEnum::BASIC;
+        return AuthorizationTypeEnum::BASIC->value;
     }
 
     /**
@@ -215,7 +215,7 @@ final class WebhookApplication extends ApplicationAbstract implements WebhookApp
      */
     public function getApplicationType(): string
     {
-        return ApplicationTypeEnum::WEBHOOK;
+        return ApplicationTypeEnum::WEBHOOK->value;
     }
 
 }
