@@ -3,22 +3,28 @@
 namespace PipesPhpSdkTests\Integration\Authorization\Base;
 
 use Exception;
+use GuzzleHttp\Psr7\Response;
+use Hanaboso\CommonsBundle\Transport\Curl\CurlManager;
 use Hanaboso\PipesPhpSdk\Application\Base\ApplicationInterface;
 use Hanaboso\PipesPhpSdk\Application\Document\ApplicationInstall;
+use Hanaboso\PipesPhpSdk\Application\Repository\ApplicationInstallRepository;
 use Hanaboso\PipesPhpSdk\Authorization\Base\OAuth1\OAuth1ApplicationInterface;
 use Hanaboso\PipesPhpSdk\Authorization\Provider\Dto\OAuth1Dto;
 use Hanaboso\PipesPhpSdk\Authorization\Provider\OAuth1Provider;
 use Hanaboso\PipesPhpSdk\Authorization\Provider\OAuth2Provider;
 use Hanaboso\Utils\Date\DateTimeUtils;
-use PipesPhpSdkTests\DatabaseTestCaseAbstract;
+use Hanaboso\Utils\String\Json;
 use PipesPhpSdkTests\Integration\Command\NullOAuth1Application;
+use PipesPhpSdkTests\KernelTestCaseAbstract;
+use PipesPhpSdkTests\MockServer\Mock;
+use PipesPhpSdkTests\MockServer\MockServer;
 
 /**
  * Class OAuth1ApplicationAbstractTest
  *
  * @package PipesPhpSdkTests\Integration\Authorization\Base
  */
-final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
+final class OAuth1ApplicationAbstractTest extends KernelTestCaseAbstract
 {
 
     /**
@@ -29,9 +35,11 @@ final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
     /**
      * @covers \Hanaboso\PipesPhpSdk\Authorization\Base\OAuth1\OAuth1ApplicationAbstract
      * @covers \Hanaboso\PipesPhpSdk\Authorization\Base\OAuth1\OAuth1ApplicationAbstract::getAuthorizationType
+     * @throws Exception
      */
     public function testGetType(): void
     {
+        $this->privateSetUp();
         self::assertEquals('oauth', $this->testApp->getAuthorizationType());
     }
 
@@ -42,8 +50,11 @@ final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
      */
     public function testIsAuthorized(): void
     {
-        $applicationInstall = $this->createApplicationInstall(
-            [ApplicationInterface::AUTHORIZATION_FORM => [OAuth1ApplicationInterface::TOKEN => '__token__']],
+        $this->privateSetUp();
+        $applicationInstall = new ApplicationInstall(
+            [
+                'settings' => [ApplicationInterface::AUTHORIZATION_FORM => [ApplicationInterface::TOKEN => '__token__']],
+            ],
         );
         self::assertTrue($this->testApp->isAuthorized($applicationInstall));
     }
@@ -55,12 +66,15 @@ final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
      */
     public function testGetApplicationForm(): void
     {
-        $applicationInstall = $this->createApplicationInstall();
+        $this->privateSetUp();
+        $applicationInstall = new ApplicationInstall();
         self::assertEquals(
             4,
-            count($this->testApp->getApplicationForms(
-                $applicationInstall,
-            )[ApplicationInterface::AUTHORIZATION_FORM][ApplicationInterface::FIELDS]),
+            count(
+                $this->testApp->getApplicationForms(
+                    $applicationInstall,
+                )[ApplicationInterface::AUTHORIZATION_FORM][ApplicationInterface::FIELDS],
+            ),
         );
     }
 
@@ -73,11 +87,13 @@ final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
      */
     public function testAuthorize(): void
     {
-        $applicationInstall = $this->createApplicationInstall(
+        $applicationInstall = new ApplicationInstall(
             [
-                ApplicationInterface::AUTHORIZATION_FORM => [
-                    OAuth1ApplicationInterface::CONSUMER_KEY    => 'key',
-                    OAuth1ApplicationInterface::CONSUMER_SECRET => 'secret',
+                'settings' => [
+                    ApplicationInterface::AUTHORIZATION_FORM => [
+                        OAuth1ApplicationInterface::CONSUMER_KEY    => 'key',
+                        OAuth1ApplicationInterface::CONSUMER_SECRET => 'secret',
+                    ],
                 ],
             ],
         );
@@ -103,7 +119,7 @@ final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
             'access_token' => '__token__',
             'expires_in'   => DateTimeUtils::getUtcDateTime('tomorrow')->getTimestamp(),
         ];
-        $applicationInstall = $this->createApplicationInstall();
+        $applicationInstall = new ApplicationInstall();
 
         $provider = $this->createPartialMock(OAuth1Provider::class, ['getAccessToken']);
         $provider->expects(self::any())->method('getAccessToken')->willReturn($token);
@@ -126,7 +142,8 @@ final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
      */
     public function testGetAndSetFrontendRedirectUrl(): void
     {
-        $applicationInstall = $this->createApplicationInstall();
+        $this->privateSetUp();
+        $applicationInstall = new ApplicationInstall();
         $this->testApp->setFrontendRedirectUrl($applicationInstall, '/redirect/url');
 
         self::assertEquals('/redirect/url', $this->testApp->getFrontendRedirectUrl($applicationInstall));
@@ -139,10 +156,31 @@ final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
      */
     public function testSaveOauthStuff(): void
     {
-        $applicationInstall = $this->createApplicationInstall();
+        $mockServer = new MockServer();
+        self::getContainer()->set('hbpf.worker-api', $mockServer);
+        $mockServer->addMock(
+            new Mock(
+                '/document/ApplicationInstall',
+                Json::decode(
+                    '[{"id":null,"user":null,"name":null,"nonEncryptedSettings":[],"encryptedSettings":"001_e\/JlNGWNzxhpMluJq6AIXqSwIGG\/kwy7UCMcQxLRg4s=:nUgeUUcCdDtl+VjR+HuO8\/YQcLVqysMzLPVogoPea14=:2PhXCdXiAaYy+EA8dptzNimqa6hkMB3x:fhdQJm5XryXFfIWfXpN894nThpk1eDLsElHN7NDAfFJqZV62kUN0E83M\/cte\/0GIfeX39nD4AHBhctVQW2rsDhfCLY83DUJgMCH95yjH83I4zWjbvFj5","settings":[],"created":"2023-02-13 13:58:17","updated":"2023-02-13 13:58:17","expires":null,"enabled":false}]',
+                ),
+                CurlManager::METHOD_POST,
+                new Response(200, [], '[]'),
+                [
+                    'created'           => '2023-02-13 13:58:17',
+                    'updated'           => '2023-02-13 13:58:17',
+                    'encryptedSettings' => '001_e/JlNGWNzxhpMluJq6AIXqSwIGG/kwy7UCMcQxLRg4s=:nUgeUUcCdDtl+VjR+HuO8/YQcLVqysMzLPVogoPea14=:2PhXCdXiAaYy+EA8dptzNimqa6hkMB3x:fhdQJm5XryXFfIWfXpN894nThpk1eDLsElHN7NDAfFJqZV62kUN0E83M/cte/0GIfeX39nD4AHBhctVQW2rsDhfCLY83DUJgMCH95yjH83I4zWjbvFj5',
+                ],
+            ),
+        );
+
+        $this->privateSetUp();
+        $applicationInstall = new ApplicationInstall();
         $callable           = $this->invokeMethod($this->testApp, 'saveOauthStuff');
         $dto                = new OAuth1Dto($applicationInstall);
-        $callable($this->dm, $dto, ['data']);
+        /** @var ApplicationInstallRepository $applicationInstallRepository */
+        $applicationInstallRepository = self::getContainer()->get('hbpf.application_install.repository');
+        $callable($applicationInstallRepository, $dto, ['data']);
 
         self::assertEquals(
             ['data'],
@@ -154,29 +192,9 @@ final class OAuth1ApplicationAbstractTest extends DatabaseTestCaseAbstract
     /**
      * @throws Exception
      */
-    protected function setUp(): void
+    protected function privateSetUp(): void
     {
-        parent::setUp();
-
         $this->testApp = self::getContainer()->get('hbpf.application.null3');
-    }
-
-    /**
-     * @param mixed[] $settings
-     *
-     * @return ApplicationInstall
-     * @throws Exception
-     */
-    private function createApplicationInstall(array $settings = []): ApplicationInstall
-    {
-        $applicationInstall = (new ApplicationInstall())
-            ->setKey('null1')
-            ->setUser('user')
-            ->setSettings($settings);
-
-        $this->pfd($applicationInstall);
-
-        return $applicationInstall;
     }
 
 }
