@@ -3,11 +3,6 @@
 namespace PipesPhpSdkTests\Integration\HbPFTableParserBundle\Handler;
 
 use Exception;
-use Hanaboso\CommonsBundle\Exception\FileStorageException;
-use Hanaboso\CommonsBundle\FileStorage\Document\File;
-use Hanaboso\CommonsBundle\FileStorage\Dto\FileContentDto;
-use Hanaboso\CommonsBundle\FileStorage\Dto\FileStorageDto;
-use Hanaboso\CommonsBundle\FileStorage\FileStorage;
 use Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandler;
 use Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandlerException;
 use Hanaboso\PipesPhpSdk\Parser\Exception\TableParserException;
@@ -36,11 +31,6 @@ final class TableParserHandlerTest extends KernelTestCaseAbstract
     private string $path;
 
     /**
-     * @var FileStorage
-     */
-    private FileStorage $storage;
-
-    /**
      * @covers \Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandler::parseToJson
      * @covers \Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandler::getFile
      *
@@ -67,74 +57,14 @@ final class TableParserHandlerTest extends KernelTestCaseAbstract
 
     /**
      * @covers \Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandler::getFile
-     * @throws Exception
-     */
-    public function testGetFile(): void
-    {
-        $file    = (new File())->setFileUrl('/url/file');
-        $storage = self::createPartialMock(FileStorage::class, ['getFileDocument', 'getFileStorage']);
-        $storage->expects(self::any())->method('getFileDocument')->willReturn($file);
-        $storage
-            ->expects(self::any())
-            ->method('getFileStorage')
-            ->willReturn(new FileStorageDto($file, $file->getFileUrl()));
-        $handler = new TableParserHandler(new TableParser(), $storage);
-        $isTmp   = FALSE;
-
-        $fileSystem = $this->createPartialMock(Filesystem::class, ['dumpFile']);
-        $fileSystem->expects(self::any())->method('dumpFile');
-
-        $path = $this->invokeMethod($handler, 'getFile', [['file_id' => '123'], &$isTmp, $fileSystem]);
-        self::assertStringContainsString('/var/www/src/HbPFTableParserBundle/Handler/', $path);
-    }
-
-    /**
-     * @covers \Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandler::getFile
-     *
-     * @throws Exception
-     */
-    public function testGetFile2(): void
-    {
-        $file    = (new File())->setFileUrl('/url/file');
-        $storage = self::createPartialMock(FileStorage::class, ['getFileDocument', 'getFileStorage']);
-        $storage->expects(self::any())->method('getFileDocument')->willReturn($file);
-        $storage
-            ->expects(self::any())
-            ->method('getFileStorage')
-            ->willThrowException(new FileStorageException());
-        $handler = new TableParserHandler(new TableParser(), $storage);
-        $isTmp   = FALSE;
-
-        self::expectException(FileStorageException::class);
-        $this->invokeMethod($handler, 'getFile', [['file_id' => '123'], &$isTmp]);
-    }
-
-    /**
-     * @covers \Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandler::getFile
      *
      * @throws Exception
      */
     public function testGetFileErr(): void
     {
-        $isTmp = FALSE;
-
         self::expectException(TableParserHandlerException::class);
         self::expectExceptionCode(TableParserHandlerException::PROPERTY_FILE_ID_NOT_SET);
-        $this->invokeMethod($this->handler, 'getFile', [[], &$isTmp]);
-    }
-
-    /**
-     * @covers \Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandler::parseToJson
-     *
-     * @throws Exception
-     */
-    public function testParseToJsonFromContent(): void
-    {
-        $content = Files::getContent(sprintf('%s/input-10.xlsx', $this->path));
-        $file    = $this->storage->saveFileFromContent(new FileContentDto($content, 'xlsx'));
-
-        $result = $this->handler->parseToJson(['file_id' => $file->getId()]);
-        self::assertEquals(Files::getContent(sprintf('%s/output-10.json', $this->path)), $result);
+        $this->invokeMethod($this->handler, 'getFile', [[], new Filesystem()]);
     }
 
     /**
@@ -182,32 +112,6 @@ final class TableParserHandlerTest extends KernelTestCaseAbstract
             ],
         );
         self::assertEquals(Files::getContent(sprintf('%s/output-10h.json', $this->path)), $result);
-        unlink($resultPath);
-    }
-
-    /**
-     * @covers \Hanaboso\PipesPhpSdk\HbPFTableParserBundle\Handler\TableParserHandler::parseFromJson
-     *
-     * @throws Exception
-     */
-    public function testParseFromJsonFromContent(): void
-    {
-        $content = Files::getContent(sprintf('%s/output-10.json', $this->path));
-        $file    = $this->storage->saveFileFromContent(new FileContentDto($content, 'json'));
-
-        $resultPath = $this->handler->parseFromJson(
-            TableParserInterface::XLSX,
-            [
-                'file_id' => $file->getId(),
-            ],
-        );
-        $result     = $this->handler->parseToJson(
-            [
-                'file_id'     => $resultPath,
-                'has_headers' => FALSE,
-            ],
-        );
-        self::assertEquals(Files::getContent(sprintf('%s/output-10.json', $this->path)), $result);
         unlink($resultPath);
     }
 
@@ -266,8 +170,7 @@ final class TableParserHandlerTest extends KernelTestCaseAbstract
     {
         parent::setUp();
 
-        $this->storage = self::getContainer()->get('hbpf.file_storage');
-        $this->handler = new TableParserHandler(new TableParser(), $this->storage);
+        $this->handler = new TableParserHandler(new TableParser());
         $this->path    = __DIR__ . '/../../Parser/data';
     }
 
